lambda/deploy-function/requirements.txt
boto3
urllib3
zipfile38
python-magic
